app/core/analyzer.py
"""
AI-based profile analysis module for Instagram followers.
This module provides intelligent analysis of follower profiles to detect potential risks.
"""

from typing import Dict, List, Optional, Any, Union
from dataclasses import dataclass
import numpy as np
import logging
import json
from pathlib import Path
import openai
from app.utils.helpers import async_retry, safe_json_loads, validate_required_fields
from app.core.logger import AuditLogger
import os
import asyncio

@dataclass
class ProfileMetrics:
    """Container for profile analysis metrics."""
    authenticity_score: float
    engagement_potential: float
    risk_level: float
    interaction_pattern: float
    account_age: float

class ProfileAnalyzer:
    """Analyzes Instagram profiles using various metrics and AI techniques."""
    
    def __init__(self, api_key: Optional[str] = None):
        """
        Initialize the analyzer with default weights and thresholds.
        
        Args:
            api_key: Optional OpenAI API key
        """
        self.weights = {
            'activity': 0.3,
            'content': 0.3,
            'interaction': 0.2,
            'age': 0.2
        }
        self.logger = logging.getLogger(__name__)
        self.audit_logger = AuditLogger()
        
        # Initialize OpenAI client
        if api_key:
            openai.api_key = api_key
        else:
            # Try to load from environment
            try:
                from dotenv import load_dotenv
                load_dotenv()
                openai.api_key = os.getenv("OPENAI_API_KEY")
            except ImportError:
                self.logger.warning("python-dotenv not installed")
        
        # Load mock data for testing
        try:
            mock_data_path = Path(__file__).parent.parent.parent / "data" / "mock_engagement.json"
            with open(mock_data_path) as f:
                self.mock_data = json.load(f)
        except Exception as e:
            self.logger.warning(f"Could not load mock data: {str(e)}")
            self.mock_data = {}
    
    @async_retry(
        max_attempts=3,
        base_delay=1.0,
        max_delay=10.0,
        exceptions=(ConnectionError, TimeoutError),
        on_retry=lambda e, attempt: logger.warning(f"Retry {attempt} due to {type(e).__name__}")
    )
    async def analyze_profile(self, profile_pic_url: str, bio: str) -> Dict[str, float]:
        """Analyze a profile picture and bio to estimate authenticity.
        
        Args:
            profile_pic_url: URL of the profile picture
            bio: User's bio text
            
        Returns:
            Dictionary of analysis metrics
        """
        try:
            # Validate inputs
            if not profile_pic_url or not bio:
                raise ValueError("Profile picture URL and bio are required")
            
            # Simulate API call delay
            await asyncio.sleep(0.5)
            
            # Return mock metrics for testing
            return {
                "authenticity_score": 0.85,
                "engagement_potential": 0.75,
                "risk_level": 0.15,
                "interaction_pattern": 0.80,
                "account_age": 0.90
            }
            
        except Exception as e:
            self.logger.error(f"Error analyzing profile: {str(e)}")
            raise
    
    @async_retry(
        max_attempts=3,
        base_delay=1.0,
        max_delay=10.0,
        exceptions=(ConnectionError, TimeoutError),
        on_retry=lambda e, attempt: logger.warning(f"Retry {attempt} due to {type(e).__name__}")
    )
    async def analyze_following_pattern(self, following: List[str]) -> Dict[str, float]:
        """Analyze following patterns to detect suspicious behavior.
        
        Args:
            following: List of usernames being followed
            
        Returns:
            Dictionary of analysis metrics
        """
        try:
            # Validate inputs
            if not following:
                raise ValueError("Following list cannot be empty")
            
            # Simulate API call delay
            await asyncio.sleep(0.5)
            
            # Return mock metrics for testing
            return {
                "following_ratio": 0.65,
                "community_connection": 0.80,
                "suspicious_pattern": 0.20
            }
            
        except Exception as e:
            self.logger.error(f"Error analyzing following pattern: {str(e)}")
            raise
    
    def get_mock_metrics(self, username: str) -> Optional[Dict[str, float]]:
        """Get mock metrics for testing.
        
        Args:
            username: Username to get metrics for
            
        Returns:
            Dictionary of mock metrics or None if not found
        """
        return self.mock_data.get(username)
    
    def analyze_profile_risk(self, profile_id: str) -> float:
        """
        Analyze a profile and return a risk score.
        
        Args:
            profile_id: The Instagram profile ID to analyze
            
        Returns:
            float: Risk score between 0 and 1
        """
        metrics = self._gather_metrics(profile_id)
        return self._calculate_risk_score(metrics)
    
    def _gather_metrics(self, profile_id: str) -> ProfileMetrics:
        """Gather various metrics about the profile."""
        # TODO: Implement actual Instagram API calls
        # For now, return dummy data
        return ProfileMetrics(
            authenticity_score=0.5,
            engagement_potential=0.6,
            risk_level=0.4,
            interaction_pattern=0.8,
            account_age=0.9
        )
    
    def _calculate_risk_score(self, metrics: ProfileMetrics) -> float:
        """Calculate overall risk score from metrics."""
        scores = [
            metrics.authenticity_score * self.weights['activity'],
            metrics.engagement_potential * self.weights['content'],
            metrics.risk_level * self.weights['interaction'],
            metrics.interaction_pattern * self.weights['interaction'],
            metrics.account_age * self.weights['age']
        ]
        return np.mean(scores) 